refactor: replace types in global types catalog closer components
