+++ client/src/components/Auth/hooks/actions/useLogoutAction.hook.ts
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom';
 import { httpClient } from '../../../../api/httpClient.api';
-import { AuthStateContext } from '../../../../types/auth.types';
+import { AuthStateContext } from '../../types/auth.types';
 
 export const useLogoutAction = (state: AuthStateContext) => {
   const navigate = useNavigate();
