+++ client/src/components/Auth/hooks/useAuthActions.hook.ts
@@ -4,7 +4,7 @@ import { useForgotPasswordAction } from './actions/useForgotPasswordAction.hook'
 import { useLogoutAction } from './actions/useLogoutAction.hook';
 import { useGoogleLoginAction } from './actions/useGoogleLoginAction.hook';
 import { useResetPasswordAction } from './actions/useResetPasswordAction.hook';
-import { AuthStateContext, AuthActions } from '../../../types/auth.types';
+import { AuthStateContext, AuthActions } from '../types/auth.types';
 
 export const useAuthActions = (state: AuthStateContext): AuthActions => {
   const login = useLoginAction(state);
