+++ client/src/components/Auth/types/auth.types.ts
@@ -1,4 +1,32 @@
-import { User } from './user.types';
+export type User = {
+  id: string;
+  _id?: string; // MongoDB ID
+  email: string;
+  username: string;
+  createdAt?: string;
+  updatedAt?: string;
+  profile?: {
+    bio?: string;
+    avatar?: string;
+    socialLinks?: {
+      github?: string;
+      linkedin?: string;
+      twitter?: string;
+    };
+  };
+  stats?: {
+    points?: number;
+    completedChallenges?: number;
+    averageScore?: number;
+    totalTimeSpent?: number;
+  };
+  settings?: {
+    theme?: 'light' | 'dark';
+    emailNotifications?: boolean;
+    pushNotifications?: boolean;
+    language?: 'pl' | 'en';
+  };
+};
 
 export type AuthState = {
   isAuthenticated: boolean;
@@ -29,4 +57,4 @@ export type AuthStateContext = {
   error: string | null;
   isAuthenticated: boolean;
   user: User | null;
-};
+};
