+++ client/src/components/Dashboard/CommunitySection/Ranking/types/ranking.types.ts
@@ -1,16 +1,45 @@
-export interface RankingUser {
-  position: number;
+export type RankingPeriod = 'daily' | 'weekly' | 'monthly' | 'allTime';
+
+export type RankingStats = {
+  rank: number;
+  rankChange: number;
+  activePlayers: number;
+  activePlayersChange: number;
+  completedChallenges: number;
+  challengesChange: number;
+  accuracy: number;
+  winStreak: number;
+  level: number;
+  points: number;
+};
+
+export type RankingUserStats = {
+  rank: number;
+  points: number;
+  level: number;
+  streak: number;
+};
+
+export type RankingBadge = {
+  id: string;
+  name: string;
+  icon: string;
+};
+
+export type RankingUser = {
+  id: string;
+  name: string;
   username: string;
-  avatar: string | null;
+  avatar?: string;
+  position: number;
+  rank: number;
+  points: number;
+  level: number;
+  badges: RankingBadge[];
+  streak: number;
+  stats: RankingStats;
   isCurrentUser?: boolean;
-  stats: {
-    level: number;
-    points: number;
-    streak: number;
-    bestStreak: number;
-    lastActive: Date | null;
-  };
-}
+};
 
 export interface RankingResponse {
   ranking: RankingUser[];
