+++ client/src/components/Dashboard/GamesSection/GamesSection.component.tsx
@@ -5,7 +5,7 @@ import { GamesList } from './components/GamesList/GamesList.component';
 import { GamesSorting } from './components/GamesSorting/GamesSorting.component';
 import { GamesSearch } from './components/GamesSearch/GamesSearch.component';
 import { GamesDifficulty } from './components/GamesDifficulty/GamesDifficulty.component';
-import { GameDifficulty } from '../../../types/games.types';
+import { GameDifficulty } from './types/games.types';
 import { Helmet } from 'react-helmet-async';
 
 export type ActiveCategory = 'all' | 'basics' | 'algorithms' | 'challenges' | 'competitions';
