+++ client/src/components/Dashboard/GamesSection/components/GameplaySection/GameplayArea/AsyncQuest/AsyncQuestStats/AsyncQuestStats.component.tsx
@@ -1,7 +1,7 @@
 import { memo } from 'react';
 import { motion } from 'framer-motion';
 import { FaCode, FaStar, FaClock } from 'react-icons/fa';
-import { GameStats } from '../../../../../types/asyncQuest.types';
+import { GameStats } from '../types/asyncQuest.types';
 import { CategoryProgress } from '../CategoryProgress/CategoryProgress.component';
 
 type AsyncQuestStatsProps = {
