+++ client/src/components/Dashboard/GamesSection/types/games.types.ts
@@ -1,24 +1,32 @@
 export type GameDifficulty = 'easy' | 'medium' | 'hard';
-
-export type ActiveCategory = 'all' | 'basics' | 'algorithms' | 'challenges' | 'competitions';
 export type SortOption = 'newest' | 'popular' | 'difficulty' | 'xp';
 
-export interface Game {
+export type Game = {
   _id: string;
   slug: string;
-  title: string;
-  description: string;
-  difficulty: GameDifficulty;
+  id: string;
   isLevelAvailable: boolean;
   requiredLevel: number;
-  rewardPoints: number;
+  rating: {
+    average: number;
+    count: number;
+  };
   completions: {
     count: number;
   };
-  createdAt: string;
+  title: string;
+  rewardPoints: number;
+  description: string;
+  category: string;
+  difficulty: GameDifficulty;
+  completedCount: number;
+  thumbnailUrl: string;
+  isCompleted: boolean;
+  xpPoints: number;
+  createdAt?: string;
   gameData?: any[];
   estimatedTime?: number;
-}
+};
 
 export interface GamesResponse {
   games: Game[];
