+++ client/src/components/Dashboard/LearnSection/types/learning.types.ts
@@ -2,19 +2,41 @@ export type LearningPath = {
   id: string;
   title: string;
   description: string;
-  difficulty: 'beginner' | 'intermediate' | 'advanced';
-  requiredLevel: number;
   estimatedTime: number;
-  lessons: Array<{
-    id: string;
-    title: string;
-  }>;
-  outcomes: string[];
-  requirements: string[];
-  progress: {
-    completed: number;
+  isLocked: boolean;
+  requiredLevel: number;
+  progress?: {
+    completed: number | any[];
     total: number;
     percentage: number;
   };
-  isLocked: boolean;
+  outcomes?: string[];
+  requirements?: string[];
+  difficulty?: 'beginner' | 'intermediate' | 'advanced';
+  tags?: string[];
+};
+
+export type Lesson = {
+  title: string;
+  slug: string;
+  description: string;
+  duration: string;
+  difficulty: 'beginner' | 'intermediate' | 'advanced';
+  category: string;
+  isCompleted: boolean;
+  progress: number;
+  points: number;
+};
+
+export type Resource = {
+  data: {
+    id: string;
+    title: string;
+    description: string;
+    url: string;
+    type: 'documentation' | 'tutorial' | 'article';
+    category: string;
+    difficulty: 'beginner' | 'intermediate' | 'advanced';
+    isRecommended: boolean;
+  };
 };
