+++ client/src/components/Dashboard/SettingsSection/SettingsContent/ProfileForm/components/UserInfoFields/UserInfoFields.component.tsx
@@ -2,7 +2,7 @@ import { memo } from 'react';
 import { FaUser, FaEnvelope } from 'react-icons/fa';
 import { FormInput } from '../../../../../../UI/Form/FormInput/FormInput.component';
 import type { UseFormRegister, FieldErrors } from 'react-hook-form';
-import type { UserProfile } from '../../../../types/settings';
+import type { UserProfile } from '../../../../types/settings.types';
 import { styles } from '../../style/ProfileForm.styles';
 
 type UserInfoFieldsProps = {
