+++ client/src/components/Dashboard/StatsSection/hooks/useStats.hook.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
-import { UserStats } from '../../../../types/stats.types';
+import { UserStats } from '../types/stats.types';
 import { useAuth } from '../../../../hooks/useAuth';
 import { fetchStats } from '../api/fetchStats.api';
