--- client/src/data/progressStats.data.ts
@@ -1,28 +1,0 @@
-import { BsCodeSquare, BsTrophy, BsClock, BsLightning } from 'react-icons/bs';
-
-export const progressStats = [
-  {
-    title: 'Ukończone Wyzwania',
-    value: '24/30',
-    Icon: BsCodeSquare,
-    description: 'Rozwiązane zadania programistyczne',
-  },
-  {
-    title: 'Zdobyte Osiągnięcia',
-    value: '12',
-    Icon: BsTrophy,
-    description: 'Odblokowane specjalne nagrody',
-  },
-  {
-    title: 'Czas Nauki',
-    value: '14h',
-    Icon: BsClock,
-    description: 'Łączny czas spędzony na platformie',
-  },
-  {
-    title: 'Najdłuższa Seria',
-    value: '7 dni',
-    Icon: BsLightning,
-    description: 'Najdłuższy streak codziennej nauki',
-  },
-];
