--- client/src/types/asyncQuest.types.ts
@@ -1,12 +1,0 @@
-export type AsyncChallenge = {
-  id: string;
-  title: string;
-  description: string;
-  code: string;
-  correct: string;
-  error: string;
-  category: 'promises' | 'async-await' | 'callbacks';
-  difficulty: 'easy' | 'medium' | 'hard';
-  points: number;
-  task: string;
-};
